# swagger.yaml
openapi: 3.0.3
info:
  title: Transaction API
  description: |
    Personal finance transaction API built with Actix-Web (Rust) and SQLite.
    Manages sources (wallets/accounts), earnings, spendings, and debts.
    App settings live in login-api as of 2026-07-27.
    
    All user-specific endpoints are scoped under `/api/user` and require a bearer JWT.
    The validated token's `sub` claim is used as the `created_by` user identifier.
  version: 2.0.0
  contact:
    name: Transaction API
servers:
  - url: http://127.0.0.1:3000
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /api/user/consumables:
    get:
      tags: [Consumables]
      summary: Get every consumable unit the caller owns
      description: >
        Units still in use come first, then the ones already used up, newest
        acquisition first.
      operationId: getConsumables
      responses:
        "200":
          description: Consumables retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseConsumables"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags: [Consumables]
      summary: Add one consumable unit
      description: >
        One call creates one unit. A multi-pack is posted as several calls,
        each with its own `consumable_id` and `unit_index`, so every unit can
        run out on its own date. Sending an id that already exists updates that
        unit, which makes a retried offline write safe.
      operationId: postConsumable
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsumableInput"
      responses:
        "200":
          description: Consumable saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseConsumable"
        "400":
          description: Missing item name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/user/consumables/{consumable_id}/out:
    put:
      tags: [Consumables]
      summary: Record the date a unit ran out
      description: >
        Sending `{"out_date": null}` puts the unit back in use, which is the
        undo for a mistaken tap.
      operationId: putConsumableOut
      parameters:
        - name: consumable_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                out_date:
                  type: string
                  nullable: true
                  example: "2026-09-19T08:30:00.000"
      responses:
        "200":
          description: Consumable updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseSuccess"
        "400":
          description: out_date could not be read as a date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "404":
          description: No consumable with that id belongs to the caller
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/user/investments:
    get:
      tags: [Investments]
      summary: Get every investment holding the caller owns
      description: >
        Grouped by kind (funds, then gold, then silver) so a client can render
        its sections without re-sorting, newest acquisition first within each.
      operationId: getInvestments
      responses:
        "200":
          description: Investments retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseInvestments"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags: [Investments]
      summary: Add or edit one investment holding
      description: >
        Sending an `investment_id` that already exists updates that holding,
        which makes a retried offline write safe. `last_unit_price` falls back
        to `buy_unit_price` when absent, so a fresh holding is worth what it
        cost until the first price update lands.
      operationId: postInvestment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvestmentInput"
      responses:
        "200":
          description: Investment saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseInvestment"
        "400":
          description: Missing name, unknown kind, or negative units/price
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/user/investments/{investment_id}/price:
    put:
      tags: [Investments]
      summary: Record a fresh valuation for one holding
      description: >
        Used for a NAB read off a broker app, or a gold/silver price pulled
        from a price API. Deliberately cannot touch `units` or
        `buy_unit_price`, so a bad price refresh can never rewrite what was
        actually bought.
      operationId: putInvestmentPrice
      parameters:
        - name: investment_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [last_unit_price]
              properties:
                last_unit_price:
                  type: number
                  format: double
                  example: 2550000
                price_source:
                  type: string
                  description: >
                    Where the price came from, e.g. `anekalogam` or `spot`.
                    Empty means it was entered by hand.
                  example: "anekalogam"
                price_updated_date:
                  type: string
                  nullable: true
                  description: Defaults to now when absent or unreadable.
                  example: "2026-08-08T09:00:00.000"
      responses:
        "200":
          description: Investment price updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseSuccess"
        "400":
          description: Negative price
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "404":
          description: No investment with that id belongs to the caller
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ResponseError"
  schemas:
...
    BaseResponse:
      type: object
      properties:
        status:
          type: string
          example: "Success"
        code:
          type: integer
          example: 200
        message:
          type: string
        description:
          type: string
        success:
          type: boolean
        data:
          nullable: true
    ResponseSuccess:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            data:
              nullable: true
              example: null
    ResponseError:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            status:
              example: "Error"
            success:
              example: false
            data:
              nullable: true
              example: null

    # ── App Settings schemas moved to login-api/swagger.yaml ────────────────

    # ── Sources ─────────────────────────────────────────────────────────────
...
    ConsumableInput:
      type: object
      required: [item_name]
      properties:
        consumable_id:
          type: string
          format: uuid
          description: Client-generated, so a unit queued offline keeps its id.
        item_name:
          type: string
          example: "Shampoo Clear 340ml"
        notes:
          type: string
          example: ""
        unit_index:
          type: integer
          default: 1
          description: Position within the batch it was bought in.
          example: 2
        unit_total:
          type: integer
          default: 1
          description: Size of that batch; 1 for a single unit.
          example: 3
        price:
          type: number
          format: double
          description: What this one unit cost.
          example: 45000
        in_date:
          type: string
          description: >
            When the unit came in. Accepts `2026-08-08T10:11:12.345`,
            `2026-08-08 10:11:12`, `2026-08-08` or an RFC 3339 stamp. Defaults
            to now when absent or unreadable.
          example: "2026-08-08T10:11:12.000"
        out_date:
          type: string
          nullable: true
          description: When it ran out. Null while the unit is still in use.
        spending_id:
          type: string
          format: uuid
          nullable: true
          description: Set when the unit was created from a transaction.
        spending_detail_id:
          type: string
          format: uuid
          nullable: true
          description: The line item it was created from.
    Consumable:
      allOf:
        - $ref: "#/components/schemas/ConsumableInput"
        - type: object
          properties:
            created_date:
              type: string
              format: date-time
            updated_date:
              type: string
              format: date-time
            created_by:
              type: string
            is_active:
              type: integer
              example: 1
    ResponseConsumable:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Consumable"
    ResponseConsumables:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Consumable"
    InvestmentInput:
      type: object
      required: [kind, name]
      description: >
        `units` and the unit prices mean whatever `kind` implies: units owned
        and NAB per unit for a fund, grams and price per gram for metal.
      properties:
        investment_id:
          type: string
          format: uuid
          description: Client-generated, so a holding queued offline keeps its id.
        kind:
          type: string
          enum: [mutual_fund, gold, silver]
          example: "gold"
        name:
          type: string
          description: Fund name, or a label such as "Antam 1gr bar".
          example: "Antam 1gr bar"
        provider:
          type: string
          description: Where it is held - Bibit, Bareksa, Pegadaian, a safe.
          example: "Pegadaian"
        units:
          type: number
          format: double
          description: Units owned for a fund, grams for metal.
          example: 5
        buy_unit_price:
          type: number
          format: double
          description: >
            Cost basis per unit. Set once and never rewritten, so gain/loss
            survives every price refresh.
          example: 2400000
        last_unit_price:
          type: number
          format: double
          nullable: true
          description: >
            Latest known value per unit. Falls back to `buy_unit_price` when
            absent or zero.
          example: 2550000
        price_source:
          type: string
          description: Where `last_unit_price` came from; empty if hand-entered.
          example: "anekalogam"
        price_updated_date:
          type: string
          nullable: true
          description: >
            When that price was recorded. Accepts the same formats as every
            other date in this API.
          example: "2026-08-08T09:00:00.000"
        notes:
          type: string
          example: ""
        acquired_date:
          type: string
          description: When it was bought. Defaults to now when absent.
          example: "2026-08-08T10:11:12.000"
    Investment:
      allOf:
        - $ref: "#/components/schemas/InvestmentInput"
        - type: object
          properties:
            created_date:
              type: string
              format: date-time
            updated_date:
              type: string
              format: date-time
            created_by:
              type: string
            is_active:
              type: integer
              example: 1
    ResponseInvestment:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Investment"
    ResponseInvestments:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Investment"
...